chore: split run inspector components
