+++ web_src/.eslint-budget-baseline.json
@@ -1,19 +1,19 @@
 {
-  "maxAllowedTotalIssues": 601,
+  "maxAllowedTotalIssues": 599,
   "maxAllowedByRule": {
-    "complexity": 212,
+    "complexity": 211,
     "@typescript-eslint/no-explicit-any": 153,
     "@typescript-eslint/no-non-null-asserted-optional-chain": 78,
-    "max-lines-per-function": 67,
-    "max-statements": 62,
+    "max-lines-per-function": 65,
+    "max-statements": 63,
     "max-lines": 15,
     "max-depth": 6,
     "max-params": 5,
     "no-eslint-disable-next-line": 3
   },
   "maxAllowedMetricMaximumByRule": {
-    "max-lines": 3851,
-    "complexity": 220
+    "max-lines": 3817,
+    "complexity": 210
   },
-  "updatedAt": "2026-07-06T14:04:36.479Z"
+  "updatedAt": "2026-07-07T21:06:50.319Z"
 }
