+++ web_src/src/ui/CanvasPage/index.runInspection.spec.tsx
@@ -1,5 +1,5 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { act, fireEvent, render as testingLibraryRender, screen, waitFor } from "@testing-library/react";
+import { act, render as testingLibraryRender, waitFor } from "@testing-library/react";
 import type { ReactElement, ReactNode } from "react";
 import { MemoryRouter } from "react-router-dom";
 import { beforeEach, describe, expect, it, vi } from "vitest";
@@ -520,148 +520,5 @@ describe("CanvasPage run inspection", () => {
 
     expect(onRunNodeDetailClose).not.toHaveBeenCalled();
     expect(selectedRunNode()?.selected).toBe(true);
-  });
-
-  it("shows the right run inspector during run inspection even when the live node inspector would be open", async () => {
-    render(
-      <MemoryRouter>
-        <CanvasPage
-          title="Canvas"
-          headerMode="version-live"
-          isRunInspectionMode
-          initialSidebar={{ isOpen: true, nodeId: "run-node-1" }}
-          runNodeDetailNodeId="run-node-1"
-          runNodeDetailCanvasId="canvas-1"
-          runNodeDetailRun={{
-            id: "run-1",
-            rootEvent: { id: "root-event-1", nodeId: "trigger-node" },
-          }}
-          nodes={[
-            {
-              id: "run-node-1",
-              position: { x: 0, y: 0 },
-              data: {
-                label: "Run node",
-                state: "success",
-                type: "component",
-              },
-            },
-          ]}
-          edges={[]}
-          buildingBlocks={[]}
-          isEditing={false}
-          activeCanvasVersionId="live-version"
-        />
-      </MemoryRouter>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByTestId("run-inspector-panel")).toBeInTheDocument();
-    });
-    expect(screen.queryByTestId("live-node-detail-pane")).not.toBeInTheDocument();
-  });
-
-  it("closes only the right run inspector when the inspector close button is clicked", async () => {
-    const onRunNodeDetailClose = vi.fn();
-    const onBackToLiveCanvas = vi.fn();
-
-    render(
-      <MemoryRouter>
-        <CanvasPage
-          title="Canvas"
-          headerMode="version-live"
-          isRunInspectionMode
-          runNodeDetailNodeId="run-node-1"
-          runNodeDetailCanvasId="canvas-1"
-          runNodeDetailRun={{
-            id: "run-1",
-            rootEvent: { id: "root-event-1", nodeId: "trigger-node" },
-          }}
-          onRunNodeDetailClose={onRunNodeDetailClose}
-          onBackToLiveCanvas={onBackToLiveCanvas}
-          nodes={[
-            {
-              id: "run-node-1",
-              position: { x: 0, y: 0 },
-              data: {
-                label: "Run node",
-                state: "success",
-                type: "component",
-              },
-            },
-          ]}
-          edges={[]}
-          buildingBlocks={[]}
-          isEditing={false}
-          activeCanvasVersionId="live-version"
-        />
-      </MemoryRouter>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByTestId("run-inspector-panel")).toBeInTheDocument();
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Close" }));
-
-    expect(onRunNodeDetailClose).toHaveBeenCalledOnce();
-    expect(onBackToLiveCanvas).not.toHaveBeenCalled();
-  });
-
-  it("shows right run inspector loading chrome while run detail is loading", () => {
-    render(
-      <MemoryRouter>
-        <CanvasPage
-          title="Canvas"
-          headerMode="version-live"
-          isRunInspectionMode
-          runCanvasLoading
-          runNodeDetailCanvasId="canvas-1"
-          nodes={[]}
-          edges={[]}
-          buildingBlocks={[]}
-          isEditing={false}
-          activeCanvasVersionId="live-version"
-        />
-      </MemoryRouter>,
-    );
-
-    expect(screen.getByRole("complementary", { name: "Run inspector" })).toBeInTheDocument();
-    expect(screen.queryByTestId("run-inspector-panel")).not.toBeInTheDocument();
-  });
-
-  it("does not open the right run inspector while editing", () => {
-    render(
-      <MemoryRouter>
-        <CanvasPage
-          title="Canvas"
-          headerMode="version-live"
-          isRunInspectionMode
-          runNodeDetailNodeId="run-node-1"
-          runNodeDetailCanvasId="canvas-1"
-          runNodeDetailRun={{
-            id: "run-1",
-            rootEvent: { id: "root-event-1", nodeId: "trigger-node" },
-          }}
-          nodes={[
-            {
-              id: "run-node-1",
-              position: { x: 0, y: 0 },
-              data: {
-                label: "Run node",
-                state: "success",
-                type: "component",
-              },
-            },
-          ]}
-          edges={[]}
-          buildingBlocks={[]}
-          isEditing
-          activeCanvasVersionId="live-version"
-        />
-      </MemoryRouter>,
-    );
-
-    expect(screen.queryByTestId("run-inspector-panel")).not.toBeInTheDocument();
   });
 });
