+++ web_src/src/ui/Runs/RunInspectorPanel.spec.tsx
@@ -2,6 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { createEvent, fireEvent, render, screen, waitFor, within } from "@testing-library/react";
 import { useState } from "react";
 import { describe, expect, it, vi, afterEach, beforeEach } from "vitest";
+import type * as ApiClient from "@/api-client";
 import type { CanvasesCanvasNodeExecution, CanvasesCanvasRun, SuperplaneComponentsNode } from "@/api-client";
 import { ThemeProvider } from "@/contexts/ThemeProvider";
 import { RunInspectorPanel } from "./RunInspectorPanel";
@@ -63,7 +64,7 @@ const listNodeQueueItemsMock = vi.fn();
 const deleteNodeQueueItemMock = vi.fn();
 
 vi.mock("@/api-client", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("@/api-client")>();
+  const actual = await importOriginal<typeof ApiClient>();
   return {
     ...actual,
     canvasesReemitTriggerEvent: (...args: unknown[]) => reemitTriggerEventMock(...args),
