+++ web_src/src/ui/Runs/RunInspectorPanel.tsx
@@ -1,25 +1,20 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import {
-  canvasesCancelExecution,
-  canvasesDeleteNodeQueueItem,
-  canvasesListNodeQueueItems,
-  canvasesReemitTriggerEvent,
   type CanvasesCanvasRun,
   type ComponentsEdge,
   type SuperplaneComponentsNode as ComponentsNode,
 } from "@/api-client";
 import { useEventExecutions } from "@/hooks/useCanvasData";
 import { appDarkModeClasses } from "@/lib/appDarkModeClasses";
-import { showErrorToast, showSuccessToast } from "@/lib/toast";
 import { cn } from "@/lib/utils";
-import { withOrganizationHeader } from "@/lib/withOrganizationHeader";
 import { RunInspectorChrome } from "./RunInspectorChrome";
 import { RunInspectorHeader } from "./RunInspectorHeader";
-import { ResizeHandle, useResizableInspectorWidth } from "./RunInspectorResize";
+import { ResizeHandle } from "./RunInspectorResize";
 import { RunInspectorStepsList } from "./RunInspectorStepsList";
 import { buildNodeMap, buildRunPresentation } from "./runPresentation";
 import { buildRunInspectorNodeSections, findRunInspectorErrorSummaries } from "./runNodeDetailModel";
+import { useResizableInspectorWidth } from "./useResizableInspectorWidth";
+import { useRunInspectorActions } from "./useRunInspectorActions";
 
 export interface RunInspectorPanelProps {
   canvasId: string;
@@ -55,96 +50,14 @@ export function RunInspectorPanel({
   );
   const errorSummaries = useMemo(() => findRunInspectorErrorSummaries(sections), [sections]);
   const inspectorWidth = useResizableInspectorWidth();
-  const queryClient = useQueryClient();
   const selectedValue = selectedNodeId ?? "";
   const [pendingErrorScrollNodeId, setPendingErrorScrollNodeId] = useState<string | null>(null);
-  const runningExecutionIds = useMemo(
-    () =>
-      sections
-        .map((section) => section.execution)
-        .filter((execution) => execution?.id && execution.state === "STATE_STARTED")
-        .map((execution) => execution!.id!),
-    [sections],
-  );
-  const stoppableNodeIds = useMemo(() => [...new Set(sections.map((section) => section.nodeId))], [sections]);
-
-  const refreshRunQueries = useCallback(async () => {
-    await queryClient.invalidateQueries({ queryKey: ["canvases"] });
-  }, [queryClient]);
-
-  const rerunMutation = useMutation({
-    mutationFn: async () => {
-      if (!run.rootEvent?.nodeId || !run.rootEvent?.id) {
-        throw new Error("Run root event is missing");
-      }
-
-      await canvasesReemitTriggerEvent(
-        withOrganizationHeader({
-          path: {
-            canvasId,
-            nodeId: run.rootEvent.nodeId,
-            eventId: run.rootEvent.id,
-          },
-        }),
-      );
-    },
-    onSuccess: async () => {
-      await refreshRunQueries();
-      showSuccessToast("Run restarted");
-    },
-    onError: (error) => {
-      console.error("Failed to restart run", error);
-      showErrorToast("Failed to restart run");
-    },
+  const actions = useRunInspectorActions({
+    canvasId,
+    run,
+    sections,
+    executionsLoading: executionsQuery.isLoading,
   });
-
-  const stopMutation = useMutation({
-    mutationFn: async () => {
-      const queuedItems = await listQueuedItemsForRun({
-        canvasId,
-        nodeIds: stoppableNodeIds,
-        rootEventId: run.rootEvent?.id,
-      });
-
-      if (runningExecutionIds.length === 0 && queuedItems.length === 0) {
-        throw new Error("No running or queued steps to stop");
-      }
-
-      await Promise.all([
-        ...runningExecutionIds.map((executionId) =>
-          canvasesCancelExecution(
-            withOrganizationHeader({
-              path: {
-                canvasId,
-                executionId,
-              },
-            }),
-          ),
-        ),
-        ...queuedItems.map((item) =>
-          canvasesDeleteNodeQueueItem(
-            withOrganizationHeader({
-              path: {
-                canvasId,
-                nodeId: item.nodeId,
-                itemId: item.itemId,
-              },
-            }),
-          ),
-        ),
-      ]);
-    },
-    onSuccess: async () => {
-      await refreshRunQueries();
-      showSuccessToast("Run stopped");
-    },
-    onError: (error) => {
-      console.error("Failed to stop run", error);
-      showErrorToast("Failed to stop run");
-    },
-  });
-  const stopActionDisabled =
-    executionsQuery.isLoading || stopMutation.isPending || (runningExecutionIds.length === 0 && !run.rootEvent?.id);
 
   const handleValueChange = (value: string) => {
     if (value) {
@@ -188,9 +101,9 @@ export function RunInspectorPanel({
         run={run}
         title={presentation.title}
         stepCount={sections.length || run.executions?.length || 0}
-        onAction={() => (presentation.status === "running" ? stopMutation.mutate() : rerunMutation.mutate())}
-        actionPending={presentation.status === "running" ? stopMutation.isPending : rerunMutation.isPending}
-        actionDisabled={presentation.status === "running" ? stopActionDisabled : !run.rootEvent?.id}
+        onAction={() => (presentation.status === "running" ? actions.stop() : actions.rerun())}
+        actionPending={presentation.status === "running" ? actions.stopPending : actions.rerunPending}
+        actionDisabled={presentation.status === "running" ? actions.stopDisabled : !run.rootEvent?.id}
       />
 
       <RunInspectorStepsList
@@ -202,42 +115,9 @@ export function RunInspectorPanel({
         componentIconMap={componentIconMap}
         onValueChange={handleValueChange}
         onJumpToError={jumpToErrorOutput}
-        onRerun={() => rerunMutation.mutate()}
-        rerunPending={rerunMutation.isPending}
+        onRerun={actions.rerun}
+        rerunPending={actions.rerunPending}
       />
     </aside>
   );
-}
-
-async function listQueuedItemsForRun({
-  canvasId,
-  nodeIds,
-  rootEventId,
-}: {
-  canvasId: string;
-  nodeIds: string[];
-  rootEventId?: string;
-}) {
-  if (!rootEventId || nodeIds.length === 0) {
-    return [];
-  }
-
-  const responses = await Promise.all(
-    nodeIds.map(async (nodeId) => {
-      const response = await canvasesListNodeQueueItems(
-        withOrganizationHeader({
-          path: { canvasId, nodeId },
-          query: { limit: 100 },
-        }),
-      );
-
-      return (
-        response.data?.items
-          ?.filter((item) => item.id && item.rootEvent?.id === rootEventId)
-          .map((item) => ({ nodeId, itemId: item.id! })) ?? []
-      );
-    }),
-  );
-
-  return responses.flat();
 }
